// Lab3/pzpi-22-8-hmyrak-mykhailo-lab3/src/types/statistics.ts
export interface LibraryStats {
  bookTitlesCount: number;
  bookItemsCount: number;
  itemsOnLoanCount: number;
  availableItemsCount: number;
  specialStorageCount: number;
  readersCount: number;
  activeReadersCount: number;
  popularBookTitlesCount: number;
  averageReadingTime: number;
  maxReadingTime: number;
}